auth-service/requirements.txt
# Django and related packages
django==5.0
djangorestframework==3.15
gunicorn
django-cors-headers

# Database and cache
psycopg2-binary
django-redis==5.2.0
redis

# Authentication and social auth
djoser==2.2.2
djangorestframework-simplejwt==5.3.1
cffi
argon2-cffi-bindings==21.2.0
PyJWT==2.8.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
rsa==4.9

# API and storage
djangorestframework-api-response==0.1.0
django-storages==1.13.1
boto3==1.24.92
botocore==1.27.92
Pillow==9.2.0

# Channels and Kafka
channels==3.0.5
channels-redis==3.4.1

# Utilities
django-environ==0.9.0
python-slugify==6.1.2
jsonpickle==3.0.1
whitenoise==6.2.0